senderApp/requirements.txt
Flask
requests
python-dotenv
psutil
pywin32; platform_system == "Windows"
gunicorn